+++ ansi_chameleon/lib/mysql_lib.py
@@ -8,7 +8,7 @@ from pymysqlreplication import BinLogStreamReader
 from pymysqlreplication.event import QueryEvent, GtidEvent, HeartbeatLogEvent
 from pymysqlreplication.row_event import DeleteRowsEvent,UpdateRowsEvent,WriteRowsEvent
 from pymysqlreplication.event import RotateEvent
-from pg_chameleon import sql_token
+from ansi_chameleon import sql_token
 from os import remove
 
 class mysql_source(object):